Finished with the solver!

+++ source/solvers/vof_curvature_projection.cc
@@ -30,9 +30,9 @@ VOFCurvatureProjection<dim>::assemble_system_matrix_and_rhs()
     update_values);
 
   // Initialize size of arrays
-  const double n_q_points =
+  const unsigned int n_q_points =
     fe_values_curvature_projection.get_quadrature().size();
-  const double n_dofs_per_cell =
+  const unsigned int n_dofs_per_cell =
     fe_values_curvature_projection.get_fe().n_dofs_per_cell();
 
   // Initialize local matrix and rhs
